imx95evk.conf: Exclude imx95-19x19-lpddr5-verdin [YOCIMX-8168]
As so many long name dtb/dtbo are added to boot partition of this consolidated machine,
it causes the vfat file number limitation:
When you create a long filename with VFAT, it uses one directory entry for the MS-DOS alias
and another entry for every 13 characters of the long filename.
In theory, a single long filename could occupy up to 21 directory entries.
The root directory has a limit of 512 files, but if you were to use the maximum length long filenames
in the root directory, you could cut this limit to a mere 24 files.

so it causes the dp_image_wic break as:
 Copying imx95-19x19-verdin-ox05b1s-isp-lt8912.dtb
 Copying imx95-19x19-verdin-panel-cap-touch-10inch-dsi.dtb
 No directory slots
 No directory slots
 Copying imx95-19x19-verdin-rm692c9.dtb
 No directory slots

Signed-off-by: Jun Zhu <junzhu@nxp.com>
(cherry picked from commit edfbd4b4b60fc7bfa6b70e69788fef16d4476809)
Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>