imx-boot: Fix dtb name failure
Now that u-boot-imx has the latest changes from upstream, there is a
build break for all 8M:
```
| NOTE: building iMX8MP - TEE=tee.bin-stmm flash_evk_stmm_capsule
| dtc -@ -I dts -O dtb -o signature.dtbo signature.dts
| fdtoverlay -i imx8mp-evk.dtb -o imx8mp-evk.dtb signature.dtbo
| Couldn't open blob from 'imx8mp-evk.dtb': No such file or directory
```

The problem is the latest imx-boot recipe does now copy the dtb from
u-boot-imx with a config-adorned name: imx8mp-evk.dtb-sd. But mkimage
is expecting imx8mp-evk.dtb. (This despite the fact that
imx8mp-evk.dtb-sd is passed in dtbs, but that's another issue for
another day.)

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>