weston-init: Fix RDP support [LF-13230]
The modification of the weston.ini command line necessary for RDP
support is missing. The problem is the weston.ini template contains
`@bindir@`, while the sed operation for the modification contains
`${bindir}`.

Fix the problem by repeating the modification in our function
`do_install:append()`, which occurs after the substitution of
`@bindir@`.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>