qtbase: Enable gbm and kms for i.MX 9 more simply
The logic for fixing the 91 and 93 builds is complex. It seems that all
that is required is to enable gbm and kms for all i.MX 9 the same as is
done for i.MX 8.

Reworks: 6c83db088e qtbase: Enable gbm and kms for mx91
Reworks: 9fddf8def6 qtbase: Add kms and gbm to 93 [LF-9315]

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>