firmware-ele-imx: Upgrade to 1.3.0 for i.MX 91 and 93 [YOCIMX-8724]
A vulnerability was identified after release in the firmware-ele-imx 0.1.2
version for i.MX 91 and 93. This commit backports the firmware-ele-imx 1.3.0
version that addresses the vulnerability for i.MX 91 and i.MX 93.

The firmware-ele-imx 1.3.0 version is not production tested and should be
used for development only for i.MX 91 and i.MX 93. Please upgrade to the
6.6.52-2.2.0 release for full production software BSP.

The vulnerability does not impact i.MX 8ULP and i.MX 95, so the original
firmware-ele-imx 0.1.2 version for the release is still marked as preferred.

Signed-off-by: Neena Busireddy <neenareddy.busireddy@nxp.com>